+++ movie-api-module/src/main/java/com/app/movie/aop/RateLimitAspect.java
@@ -1,6 +1,7 @@
 package com.app.movie.aop;
 
 import com.app.movie.application.RateLimiterRedisService;
+import com.app.movie.filter.ClientIpHolder;
 import com.app.movie.presentation.TooManyRequestException;
 import jakarta.servlet.http.HttpServletRequest;
 import org.aspectj.lang.annotation.Aspect;
