+++ movie-api-module/src/main/java/com/app/movie/filter/ClientIpHolder.java
@@ -1,4 +1,4 @@
-package com.app.movie.aop;
+package com.app.movie.filter;
 
 
 public class ClientIpHolder {
